tools/firecrawl/requirements.txt
firecrawl-py
python-dotenv
fastapi
uvicorn